move the locators extension
