+++ ui/user/src/tests/vitest-browser.d.ts
@@ -1,0 +1,7 @@
+import type { Locator } from 'vitest/browser';
+
+declare module 'vitest/browser' {
+	interface LocatorSelectors {
+		locator(selector: string): Locator;
+	}
+}
